Able to add articles to database and have them render to the DOM. Also added timestamp to saved article info
