+++ src/components/news/News.js
@@ -1,6 +1,7 @@
 import React from 'react'
-import { Container, Button, Modal, ModalHeader, ModalBody, ModalFooter, Form, FormGroup, Label, Input } from 'reactstrap'
+import { Container } from 'reactstrap'
 import NewsList from './NewsList'
+import NewsModal from './NewsModal'
 import API from './../../modules/API/API'
 
 export default class News extends React.Component {
@@ -9,19 +10,28 @@ export default class News extends React.Component {
     news: []
   }
 
+  // Fetches all articles from the database and adds them to state.
   newsLog = () => {
     let newState = {}
-
-    API.getData("news")
+    return API.getData("news")
       .then(news => newState.news = news)
       .then(() => this.setState(newState))
   }
 
+  saveArticle = (articleInfo) => {
+    API.saveData("news", articleInfo)
+      .then(() => {
+      // Clears main container and pulls new news dashboard with additional article.
+      return this.newsLog()
+    })
+  }
+
+  // OnClick functionality to delete saved articles
   handleDelete = (id) => {
     API.deleteData("news", id)
       .then(() => API.getData("news"))
       .then((news) =>
-      this.setState({news: news}))
+        this.setState({ news: news }))
   }
 
   componentDidMount() {
@@ -32,62 +42,8 @@ export default class News extends React.Component {
     return (
       <Container>
         <h1 className="text-center mt-5">News Around the 'Berg!</h1>
-        <NewsModal />
+        <NewsModal saveArticle={this.saveArticle} />
         <NewsList news={this.state.news} handleDelete={this.handleDelete} />
       </Container>)
-  }
-}
-class NewsModal extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      modal: false
-    };
-
-    this.toggle = this.toggle.bind(this);
-  }
-
-  toggle() {
-    this.setState({
-      modal: !this.state.modal
-    });
-  }
-
-    handleFieldChange = evt => {
-    const stateToChange = {}
-    stateToChange[evt.target.id] = evt.target.value
-    this.setState(stateToChange)
-    console.log(stateToChange)
-  }
-
-  render() {
-    return (
-      <div className="text-center m-3">
-        <Button className="center" color="success" onClick={this.toggle}>Add Article</Button>
-        <Modal isOpen={this.state.modal} toggle={this.toggle} className={this.props.className}>
-          <ModalHeader toggle={this.toggle}>Add Article</ModalHeader>
-          <ModalBody>
-            <Form>
-              <FormGroup>
-                <Label for="articleTitle">Title</Label>
-                <Input onChange={this.handleFieldChange} type="text" name="title" id="title" placeholder="Title of article" />
-              </FormGroup>
-              <FormGroup>
-                <Label for="articleURL">URL</Label>
-                <Input onChange={this.handleFieldChange} type="text" name="url" id="url" placeholder="Link to article" />
-              </FormGroup>
-              <FormGroup>
-                <Label for="articleSummary">Article Summary</Label>
-                <Input onChange={this.handleFieldChange} type="textarea" name="text" id="summary" placeholder="A brief summary of the article" />
-              </FormGroup>
-            </Form>
-          </ModalBody>
-          <ModalFooter>
-            <Button color="primary" onClick={this.toggle}>Save</Button>{' '}
-            <Button color="secondary" onClick={this.toggle}>Cancel</Button>
-          </ModalFooter>
-        </Modal>
-      </div>
-    );
   }
 }
