+++ src/components/news/NewsList.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Col, Row, ListGroup, ListGroupItem, Badge, Button } from 'reactstrap'
+import moment from 'moment'
 
 // Displays saved articles
 export default class NewsList extends React.Component {
@@ -14,6 +15,7 @@ export default class NewsList extends React.Component {
                 <Col xs="3" className="d-flex align-items-center">
                   <Badge><i className="icon-book-open "></i></Badge></Col>
                 <Col mx="auto"><a style={{ fontSize: 20 }} href={article.url}>{article.title}</a>
+                  <p>{moment(article.timestamp).format("MM-DD-YYYY")}</p>
                   <p>{article.summary}</p>
                 </Col>
                 <Col xs="3" className="d-flex align-items-center">
